Finished migrating from Scores to Results type
